# docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateInventoryBatchRequest:
    properties:
      cgst_rate:
        description: Tax Configuration
        maximum: 100
        minimum: 0
        type: number
      cost_price:
        type: number
      custom_tax_ids:
        items:
          type: string
        type: array
      expiry_date:
        type: string
      is_tax_exempt:
        type: boolean
      quantity:
        type: integer
      sgst_rate:
        maximum: 100
        minimum: 0
        type: number
      variant_id:
        type: string
      warehouse_id:
        type: string
    required:
    - cost_price
    - expiry_date
    - quantity
    - variant_id
    - warehouse_id
    type: object
  models.CreateInventoryTransactionRequest:
    properties:
      note:
        type: string
      quantity_change:
        type: integer
      related_entity_id:
        type: string
      transaction_type:
        type: string
    required:
    - quantity_change
    - transaction_type
    type: object
...
  models.CreateProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  models.InventoryBatchResponse:
    properties:
      cgst_rate:
        description: Tax Configuration
        type: number
      cost_price:
        type: number
      created_at:
        type: string
      custom_tax_ids:
        items:
          type: string
        type: array
      expiry_date:
        type: string
      id:
        type: string
      is_tax_exempt:
        type: boolean
      sgst_rate:
        type: number
      total_quantity:
        type: integer
      updated_at:
        type: string
      variant_id:
        type: string
      warehouse_id:
        type: string
    type: object
  models.InventoryTransactionResponse:
    properties:
      batch_id:
        type: string
      id:
        type: string
      note:
        type: string
      occurred_at:
        type: string
      performed_by:
        type: string
      quantity_change:
        type: integer
      related_entity_id:
        type: string
      transaction_type:
        type: string
    type: object
...
  models.ProductResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  utils.ErrorResponseModel:
    properties:
      error:
        example: Detailed error message
        type: string
      message:
        example: Error occurred
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: info@kisanlink.in
    name: Kisanlink Support
    url: https://github.com/Kisanlink/fpo-erp
  description: Comprehensive ERP system for agricultural cooperatives with multi-tenant
    architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Kisanlink ERP API
  version: "1.0"
paths:
...
  /api/v1/batches:
    post:
      consumes:
      - application/json
      description: Create a new inventory batch (requires authentication)
      parameters:
      - description: Batch data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateInventoryBatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Batch created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.InventoryBatchResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Create Inventory Batch
      tags:
      - Inventory
  /api/v1/batches/{id}:
    get:
      description: Retrieve a specific inventory batch by ID
      parameters:
      - description: Batch ID
        example: BTCH_12345678
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Batch retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.InventoryBatchResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "404":
          description: Batch not found
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Inventory Batch
      tags:
      - Inventory
  /api/v1/batches/{id}/transactions:
    get:
      description: Retrieve all inventory transactions for a specific batch
      parameters:
      - description: Batch ID
        example: BTCH_12345678
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transactions retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.InventoryTransactionResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Transactions by Batch
      tags:
      - Inventory
    post:
      consumes:
      - application/json
      description: Create a new inventory transaction for a batch (requires authentication)
      parameters:
      - description: Batch ID
        example: BTCH_12345678
        in: path
        name: id
        required: true
        type: string
      - description: Transaction data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateInventoryTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.InventoryTransactionResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Create Inventory Transaction
      tags:
      - Inventory
  /api/v1/batches/expiring:
    get:
      description: Retrieve inventory batches that are expiring within specified days
      parameters:
      - description: 'Number of days to check (default: 30)'
        example: 30
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Expiring batches retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.InventoryBatchResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Expiring Batches
      tags:
      - Inventory
  /api/v1/batches/low-stock:
    get:
      description: Retrieve inventory batches with stock below threshold
      parameters:
      - description: 'Stock threshold (default: 10)'
        example: 10
        in: query
        name: threshold
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Low stock batches retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.InventoryBatchResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Low Stock Batches
      tags:
      - Inventory
...
  /api/v1/products:
    get:
      description: Retrieve all products (requires authentication)
      produces:
      - application/json
      responses:
        "200":
          description: Products retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ProductResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get All Products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product (requires authentication)
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ProductResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Create Product
      tags:
      - Products
...
  /api/v1/products/availability:
    get:
      description: Retrieve availability information for all products across warehouses
      produces:
      - application/json
      responses:
        "200":
          description: Products availability retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ProductAvailabilityResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Products Availability
      tags:
      - Inventory
...
  /api/v1/variants/{id}/batches:
    get:
      description: Retrieve all inventory batches for a specific product variant
      parameters:
      - description: Variant ID
        example: PVAR_12345678
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Batches retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.InventoryBatchResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "403":
          description: Forbidden - insufficient permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "409":
          description: Conflict - resource already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "422":
          description: Unprocessable Entity - validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseModel'
      security:
      - BearerAuth: []
      summary: Get Batches by Variant
      tags:
      - Inventory
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"